Treatment Stratification Hotfix (#110)
* removing the treatment stratification in svr

* adding db schema and sim.conf to scripts and removing old python files

* Updating treatment initiation

* moving files to examples folder per PR comment

* quick update to ensure the correct callback function is called

* Huge catch for checking if the person is alive before attempting to clear

* Fixing omp, unnecessary logging, and simplifying ValidExecute

+++ src/model/internals/person_internals.hpp
@@ -4,7 +4,7 @@
 // Created Date: 2025-04-18                                                  //
 // Author: Matthew Carroll                                                    //
 // -----                                                                      //
-// Last Modified: 2025-05-14                                                  //
+// Last Modified: 2025-06-10                                                  //
 // Modified By: Matthew Carroll                                               //
 // -----                                                                      //
 // Copyright (c) 2025 Syndemics Lab at Boston Medical Center                  //
@@ -274,7 +274,7 @@ private:
     const std::string _log_name;
 
     size_t _id;
-    size_t _current_time = 0;
+    int _current_time = 0;
 
     data::Sex _sex = data::Sex::kMale;
     int _age = 0;
